% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-calcs.R
\name{get_std_curve_value}
\alias{get_std_curve_value}
\title{Function to calculate standard curve}
\usage{
get_std_curve_value(std_fluorescence, std_gfp)
}
\arguments{
\item{std_fluorescence}{Fluorescence for the standard curve.}

\item{std_gfp}{GFP for the standard curve.}
}
\description{
Function to calculate standard curve
}
