% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_gfp_utils.R
\name{plot_bar_fluorescence}
\alias{plot_bar_fluorescence}
\title{Plot fluorescence}
\usage{
plot_bar_fluorescence(mat_sample_fluorescence)
}
\arguments{
\item{mat_sample_fluorescence}{Input data.}
}
\description{
Plot fluorescence
}
