% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_bca_utils.R
\name{plot_bar_prot}
\alias{plot_bar_prot}
\title{Plot GFP}
\usage{
plot_bar_prot(
  df_with_pred_gfp,
  wildtype_sample,
  var_to_plot = "GFP (g/kg)",
  ylab = "GFP (g/kg)",
  percent_lab = "Percentage of GFP"
)
}
\arguments{
\item{df_with_pred_gfp}{Data frame with predicted GFP.}

\item{wildtype_sample}{Name of wildtype/control sample.}

\item{var_to_plot}{Variable name to plot.}

\item{ylab}{Y-axis name.}

\item{percent_lab}{Secondary Y-axis name.}
}
\description{
Plot GFP
}
