% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_gfp_utils.R
\name{predict_gfp_from_fluorescence}
\alias{predict_gfp_from_fluorescence}
\title{Function to get GFP value if fluorescence known}
\usage{
predict_gfp_from_fluorescence(df_tidied, fit)
}
\arguments{
\item{df_tidied}{Tidied fata frame.}

\item{fit}{Fit object.}
}
\description{
Function to get GFP value if fluorescence known
}
