% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_gfp_utils.R
\name{get_fluorescence_input}
\alias{get_fluorescence_input}
\title{Function to get the value, given input data}
\usage{
get_fluorescence_input(mdata)
}
\arguments{
\item{mdata}{Sample fluorescence matrix.}
}
\description{
Function to get the value, given input data
}
