% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-plots.R
\name{plot_bar_fluorescence}
\alias{plot_bar_fluorescence}
\title{Plot fluorescence}
\usage{
plot_bar_fluorescence(mat_sample_fluorescence, ylab = "Flouresence")
}
\arguments{
\item{mat_sample_fluorescence}{Input data.}

\item{ylab}{Y-axis name.}
}
\description{
Plot fluorescence
}
