% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-plots.R
\name{plot_std_curve_and_pred}
\alias{plot_std_curve_and_pred}
\title{Plot standard curve and predictions}
\usage{
plot_std_curve_and_pred(
  df_std_curve,
  df_with_pred_gfp,
  fit,
  xlab = "GFP (ng)",
  ylab = "Flouresence"
)
}
\arguments{
\item{df_std_curve}{Standard curve data frame.}

\item{df_with_pred_gfp}{Data frame with predicted GFP.}

\item{fit}{Fit object.}

\item{xlab}{X-axis name.}

\item{ylab}{Y-axis name.}
}
\description{
Plot standard curve and predictions
}
